+++ packages/config/src/doc/IProfInfoProps.ts
@@ -13,17 +13,17 @@
  * Options that will affect the behavior of the ProfileInfo class.
  * They are supplied on the ProfileInfo getAllProfiles method.
  */
- export interface IGetAllProfilesOptions {
+export interface IGetAllProfilesOptions {
 
-  /**
+    /**
    * Indicates if the home directory should be excluded
    * The default is false.
    */
-   excludeHomeDir?: boolean;
+    excludeHomeDir?: boolean;
 
-  /**
+    /**
    * Implements a custom pattern of folders that should be excluded
    * from the getAllProfiles method call.
    */
-  // excludePattern?: string;
+    // excludePattern?: string;
 }